+++ src/main/java/com/example/domain/service/WeightService.java
@@ -12,10 +12,10 @@ public interface WeightService {
 	public void recordWeight(WeightRecord record, Authentication authentication);
 	
 	//今までの最低体重を取得
-	public Integer getMinBodyWeight(int userId);
+	public Double getMinBodyWeight(int userId);
 	
 	//今までの最大体重を取得
-	public Integer getMaxBodyWeight(int userId);
+	public Double getMaxBodyWeight(int userId);
 	
 	//直近7日間の体重を取得
 	public List<WeightRecord> getBodyWeightForLast7Days(int userId, int size, int offset);
